Remove arg name/value support
arg name/value support is added in the very early stage to be able
to substitube name specified in text code to the value. E.g.,
it can be used to provide the address of format string for printf
function.

arg name/value support is simple to implement and solve real problems.
But it's functionality can be replace by 'R_386_32' relocation entry.
Remove it to avoid redundancy and simplify API.